+++ lib/mix/lib/mix/tasks/help.ex
@@ -31,7 +31,7 @@ defmodule Mix.Tasks.Help do
       max(size(task), acc)
     end
 
-    sorted = Enum.qsort(docs)
+    sorted = Enum.sort(docs)
 
     Enum.each sorted, fn({ task, doc }) ->
       shell.info format('mix ~-#{max}s # ~s', [task, doc])