+++ lib/mix/lib/mix/tasks/local.ex
@@ -19,7 +19,7 @@ defmodule Mix.Tasks.Local do
       max(size(task), acc)
     end
 
-    sorted = Enum.qsort(docs)
+    sorted = Enum.sort(docs)
 
     Enum.each sorted, fn({ task, doc }) ->
       shell.info format('mix ~-#{max}s # ~s', [task, doc])